LayoutTests/imported/w3c/web-platform-tests/scroll-to-text-fragment/scroll-to-text-fragment-service-worker.https.html

<!DOCTYPE html>
<title>Service workers text fragment directive visibility</title>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/common/utils.js"></script>
<script src="stash.js"></script>
<body>
<script>

async function registerServiceWorker()
{
    const registration = await navigator.serviceWorker.register("resources/service-worker.js", { scope : "resources/" });
    let activeWorker = registration.active;
    if (activeWorker)
        return registration;
    activeWorker = registration.installing;
    return new Promise(resolve => {
        activeWorker.addEventListener('statechange', () => {
            if (activeWorker.state === "activated")
                resolve(registration);
        });
    });
}

promise_test(async (test) => {
    registration = await registerServiceWorker();
}, "Setup worker");

promise_test(t => new Promise((resolve, reject) => {
  test_driver.bless('Open a URL with a text fragment directive', () => {
    window.addEventListener('message', (event) => {

        assert_equals(event.data.indexOf(':~:'), -1, 'Expected fragment directive to be stripped from the URL.');
        resolve()

    }, false);
    window.open(`resources/check-with-service-worker.html?#before_directive:~:text=should_not_be_visible_to_js`);
  });
}), `Test navigation with service worker.`);

</script>
</body>
